Create 002-Function_to_determine_if_youre_subject_to_mandatory_voting_under_Brazilian_law
Brazil is among a minority of countries where voting in elections is compulsory, but voting is optional for those under 18 or those 65 and over. This code asks the user his age and informs him whether he is or would be obliged to vote under Brazilian law, based on his age.